# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-Vendor Service Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseErrorsItem = {
  field: string;
  message: string;
};

export interface ErrorResponse {
  success: boolean;
  message: string;
  errors?: ErrorResponseErrorsItem[];
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface VendorRegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  businessName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  success: boolean;
  token: string;
  user: UserProfile;
}

export interface Pagination {
  page: number;
  limit: number;
  total: number;
  pages: number;
}

export interface UserListResponse {
  success: boolean;
  data: UserProfile[];
  pagination: Pagination;
}

export interface Vendor {
  id: string;
  userId: string;
  businessName: string;
  isApproved: boolean;
  createdAt: string;
}

export interface VendorListResponse {
  success: boolean;
  data: Vendor[];
  pagination: Pagination;
}

export type VendorDashboardData = {
  totalServices: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  totalEarnings: number;
};

export interface VendorDashboard {
  success: boolean;
  data: VendorDashboardData;
}

export interface Service {
  id: string;
  title: string;
  description: string;
  price: number;
  category: string;
  vendorId: string;
  rating: number;
  reviewCount: number;
  createdAt: string;
}

export type ServiceDetailVendor = {
  id: string;
  businessName: string;
};

export interface ServiceDetail {
  id: string;
  title: string;
  description: string;
  price: number;
  category: string;
  vendorId: string;
  vendor: ServiceDetailVendor;
  rating: number;
  reviewCount: number;
  createdAt: string;
}

export interface ServiceListResponse {
  success: boolean;
  data: Service[];
  pagination: Pagination;
}

export interface CreateServiceInput {
  title: string;
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
}

export interface UpdateServiceInput {
  title?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  category?: string;
}

export interface Booking {
  id: string;
  userId: string;
  vendorId: string;
  serviceId: string;
  date: string;
  time: string;
  status: string;
  paymentStatus: string;
  createdAt: string;
}

export interface BookingListResponse {
  success: boolean;
  data: Booking[];
  pagination: Pagination;
}

export interface CreateBookingInput {
  serviceId: string;
  date: string;
  time: string;
}

export type UpdateBookingStatusInputStatus =
  (typeof UpdateBookingStatusInputStatus)[keyof typeof UpdateBookingStatusInputStatus];

export const UpdateBookingStatusInputStatus = {
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusInput {
  status: UpdateBookingStatusInputStatus;
}

export type ReviewUser = {
  name: string;
};

export interface Review {
  id: string;
  userId: string;
  serviceId: string;
  rating: number;
  comment: string;
  user: ReviewUser;
  createdAt: string;
}

export interface ReviewListResponse {
  success: boolean;
  data: Review[];
  pagination: Pagination;
}

export interface CreateReviewInput {
  serviceId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment: string;
}

export interface PaymentInput {
  bookingId: string;
  amount: number;
}

export interface PaymentResponse {
  success: boolean;
  message: string;
  transactionId: string;
  bookingId: string;
  amount: number;
  paymentStatus: string;
}

export interface ApproveVendorInput {
  isApproved: boolean;
}

export type PlatformReportDataBookingsByStatus = {
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
};

export type PlatformReportData = {
  totalUsers: number;
  totalVendors: number;
  approvedVendors: number;
  pendingVendors: number;
  totalServices: number;
  totalBookings: number;
  bookingsByStatus: PlatformReportDataBookingsByStatus;
  totalRevenue: number;
};

export interface PlatformReport {
  success: boolean;
  data: PlatformReportData;
}

export type GetVendorBookingsParams = {
  status?: string;
  page?: number;
  limit?: number;
};

export type GetServicesParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  search?: string;
  sortBy?: GetServicesSortBy;
  sortOrder?: GetServicesSortOrder;
  page?: number;
  limit?: number;
};

export type GetServicesSortBy =
  (typeof GetServicesSortBy)[keyof typeof GetServicesSortBy];

export const GetServicesSortBy = {
  price: "price",
  rating: "rating",
  createdAt: "createdAt",
} as const;

export type GetServicesSortOrder =
  (typeof GetServicesSortOrder)[keyof typeof GetServicesSortOrder];

export const GetServicesSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;

export type GetMyBookingsParams = {
  status?: string;
  page?: number;
  limit?: number;
};

export type GetServiceReviewsParams = {
  page?: number;
  limit?: number;
};

export type AdminGetUsersParams = {
  page?: number;
  limit?: number;
  role?: string;
};

export type AdminGetVendorsParams = {
  page?: number;
  limit?: number;
  approved?: boolean;
};

export type AdminGetBookingsParams = {
  page?: number;
  limit?: number;
  status?: string;
};