# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Vendor Service Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register as customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Add a review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/service/{serviceId}:
    get:
      operationId: getServiceReviews
      tags: [reviews]
      summary: Get reviews for a service
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/vendors:
    get:
      operationId: adminGetVendors
      tags: [admin]
      summary: Get all vendors
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: approved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
  /admin/vendors/{id}/approve:
    patch:
      operationId: adminApproveVendor
      tags: [admin]
      summary: Approve or reject vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveVendorInput"
      responses:
        "200":
          description: Vendor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/bookings:
    get:
      operationId: adminGetBookings
      tags: [admin]
      summary: Get all bookings
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
  /admin/reports:
    get:
      operationId: adminGetReports
      tags: [admin]
      summary: Get platform analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformReport"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required: [field, message]
      required: [success, message]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [success, token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
    Vendor:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, businessName, isApproved, createdAt]
    VendorListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vendor"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
...
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vendorId:
          type: string
        serviceId:
          type: string
        date:
          type: string
        time:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, vendorId, serviceId, date, time, status, paymentStatus, createdAt]
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
...
    Review:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceId:
          type: string
        rating:
          type: number
        comment:
          type: string
        user:
          type: object
          properties:
            name:
              type: string
          required: [name]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, serviceId, rating, comment, user, createdAt]
    ReviewListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [success, data, pagination]
    CreateReviewInput:
      type: object
      properties:
        serviceId:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [serviceId, rating, comment]
...
    ApproveVendorInput:
      type: object
      properties:
        isApproved:
          type: boolean
      required: [isApproved]
    PlatformReport:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalUsers:
              type: integer
            totalVendors:
              type: integer
            approvedVendors:
              type: integer
            pendingVendors:
              type: integer
            totalServices:
              type: integer
            totalBookings:
              type: integer
            bookingsByStatus:
              type: object
              properties:
                pending:
                  type: integer
                confirmed:
                  type: integer
                completed:
                  type: integer
                cancelled:
                  type: integer
              required: [pending, confirmed, completed, cancelled]
            totalRevenue:
              type: number
          required: [totalUsers, totalVendors, approvedVendors, pendingVendors, totalServices, totalBookings, bookingsByStatus, totalRevenue]
      required: [success, data]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
      required: [page, limit, total, pages]